Add arena countdown/start and auto-join system
Implement timed game start and automatic joining: Arena now tracks players, uses a countdown task and STARTING state to trigger GameManager.startGame. ArenaManager gained autoJoin logic, arena loading from arenas.yml, and helpers to read goals/spawn points; arenas are stored in a list and constructed with GameManager, plugin and language manager. Refactor types and constructors: Goal and SpawnPoint coordinates changed to double; BallArena and Arena constructors updated to accept GameManager/Plugin/LanguageManager. Commands refactored: JoinGameCommand renamed to GameCommand (base), new FootballCommand created and JoinFootballCommand removed; SulfurFun now wires GameManager into ArenaManager and registers updated command names. Configuration and language resources updated (countdownSeconds, new language keys) and an example arenas.yml added.

+++ src/main/java/de/jgaertig/sulfurFun/arena/Goal.java
@@ -2,9 +2,9 @@ package de.jgaertig.sulfurFun.arena;
 
 public class Goal {
 
-    int corner_1_x, corner_1_y, corner_1_z, corner_2_x, corner_2_y, corner_2_z;
+    double corner_1_x, corner_1_y, corner_1_z, corner_2_x, corner_2_y, corner_2_z;
 
-    public Goal(int corner_1_x, int corner_1_y, int corner_1_z, int corner_2_x, int corner_2_y, int corner_2_z) {
+    public Goal(double corner_1_x, double corner_1_y, double corner_1_z, double corner_2_x, double corner_2_y, double corner_2_z) {
         this.corner_1_x = corner_1_x;
         this.corner_1_y = corner_1_y;
         this.corner_1_z = corner_1_z;
